+++ backend/dbSchemas.py
@@ -84,3 +84,4 @@ class AnswerUpdate(BaseModel):
     task: Optional[str] = None
     task: Optional[str] = None
     chatId: Optional[int] = None
+
