+++ backend/main.py
@@ -4,7 +4,7 @@ from fastapi.middleware.cors import CORSMiddleware
 from db import Base, engine
 from routers import chats, paragraphs, projects, answers
 from sourceDocument import generateSourceDocument
-from ai import aiChat, get_models
+from ai import aiChat, getModels
 
 
 Base.metadata.create_all(bind=engine)
@@ -27,7 +27,7 @@ app.include_router(answers.router, prefix="/answers", tags=["answers"])
 
 app.add_api_route("/aiChat", aiChat, methods=["POST", "PUT"], tags=["ai"])
 
-app.add_api_route("/aimodels", get_models, methods=["GET"], tags=["ai"])
+app.add_api_route("/aimodels", getModels, methods=["GET"], tags=["ai"])
 
 app.add_api_route(
     "/promptverzeichnis",
